refactor: Modify Unit Test
- Merge Decoding Test (ExpoInformationTest + KoreaItemTest)
- Modify initialization of NSDataAsset
- Modify Optional set (Forced Unwrapping -> Optional Chaining)
